Move the notification helpers out of the services barrel
services/index.ts opens by importing the crypto and auth layers, so any
component reaching it for a toast pulls WASM into its chunk. The delete
confirmations and the transfer row had deliberately avoided that barrel
until reporting an error required it.

They now live in services/notify.ts, which needs only i18n and the
notification library. The barrel re-exports them, so every existing
caller is untouched.
